refactor(company-monitoring): extract the Stage 0 engine and close the follow-ups
Clears the four items deferred from the review of #6186.

Engine extraction (one implementation, reachable by production):
- shared/company-monitoring-evaluation.ts now holds the statistics, canonical
  JSON, schema pinning and gate logic. Nothing was reachable outside the test
  before, so any future dark contract would have re-derived the rules from prose.
- evaluateStage0 takes the approved digest as an argument. The digest literal,
  the recorded STOP reasons and the frozen floors stay in the test source, so
  weakening a gate still cannot be done by editing the engine alone -- and
  viability.md's claim that the digest is compared against a literal held
  outside the frozen protocol stays true.

Operator affordance:
- scripts/recompute-company-monitoring-digest.mjs prints the digest and names
  the two approval sites to update; --check verifies fixture, test literal and
  recomputation agree. Previously the only way to obtain a new digest was to
  provoke an AssertionError and scrape its diff. It deliberately writes neither
  site: that is a product-owner decision, not a keystroke.

Typechecking (this is what npm run typecheck never covered):
- tsconfig.contract-tests.json + typecheck:contract-tests, wired into the
  EXISTING required typecheck job so no branch-protection context changes.
- It immediately found real defects, not just unused imports: evaluateAdmissionMetric
  computed its lower bound from the unvalidated result.numerator rather than the
  range-checked local, percentileOrderStatistic could return undefined into an
  exact-equality assertion, and several index accesses were unguarded.

Admission quality:
- evaluateAdmissionQuality is a real caller for evaluateAdmissionMetric, which
  had none, leaving the frozen metric table inert. It is deliberately NOT part of
  evaluateStage0: every admission metric is defined over published decisions or a
  Stage 3/4 blind corpus, so folding them in would make Stage 0 unsatisfiable
  rather than stricter. Unmeasured metrics and unrecognised measurements are both
  stop reasons, and a test pins the boundary that a Stage 0 `continue` says
  nothing about admission quality.

The fixture and the approved threshold digest are unchanged. 32 -> 33 tests,
typecheck and biome clean, and a re-run two-sided mutation sweep across both
files kills all 21 mutants with the no-op control surviving.

Found by code review of #6186.
