+++ .github/workflows/typecheck.yml
@@ -49,6 +49,10 @@ jobs:
       - run: npm ci
       - run: npm run typecheck
       - run: npm run typecheck:api
+      # `npm run typecheck` covers src/ only, so tests/ and shared modules reached only from tests
+      # were never type-checked by any job. Runs inside the existing required `typecheck` context
+      # rather than as a new job, so no branch-protection context list has to change.
+      - run: npm run typecheck:contract-tests
       # Mirror the pre-push convex typecheck so direct pushes to main
       # and `--no-verify` bypasses can't land stale convex/_generated
       # types. See .husky/pre-push for the rationale.
